requirements.txt
leancloud-sdk
djangorestframework
django>=1.10.0
